\name{PurBayes-package}
\alias{PurBayes-package}

\docType{package}
\title{Bayesian Estimation of Tumor Purity and Clonality}
\description{
PurBayes is an MCMC-based mixture-model algorithm that uses next-generation sequencing data to estimate tumor purity and clonality for paired tumor-normal data.

NOTE: Requires JAGS to be installed.}

\details{
\tabular{ll}{
Package: \tab PurBayes\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-05-13\cr
License: \tab GPL-2\cr
}
}
\author{Nicholas B. Larson

Maintainer: Nicholas B. Larson <larson.nicholas@mayo.edu>
}

